local_server/requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.10.4
openai==1.57.0
